--- src/main/java/alcoholboot/toastit/feature/climatecocktail/repository/RecommendByWeatherRepository.java
@@ -1,10 +1,0 @@
-package alcoholboot.toastit.feature.climatecocktail.repository;
-
-import alcoholboot.toastit.feature.climatecocktail.entity.RecommendByWeatherEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RecommendByWeatherRepository extends JpaRepository<RecommendByWeatherEntity, Long> {
-    RecommendByWeatherEntity findByWeather(String weather);
-}
