+++ src/main/java/alcoholboot/toastit/feature/weathercocktail/entity/WeatherCocktailEntity.java
@@ -1,4 +1,4 @@
-package alcoholboot.toastit.feature.climatecocktail.entity;
+package alcoholboot.toastit.feature.weathercocktail.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Table(name = "weather_cocktails")
 @Getter
 @Setter
-public class RecommendByWeatherEntity {
+public class WeatherCocktailEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
