+++ src/main/java/alcoholboot/toastit/feature/weathercocktail/entity/WeatherEntity.java
@@ -1,4 +1,4 @@
-package alcoholboot.toastit.feature.climatecocktail.entity;
+package alcoholboot.toastit.feature.weathercocktail.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
