+++ src/main/java/alcoholboot/toastit/feature/weathercocktail/repository/WeatherRepository.java
@@ -1,7 +1,7 @@
-package alcoholboot.toastit.feature.climatecocktail.repository;
+package alcoholboot.toastit.feature.weathercocktail.repository;
 
-import alcoholboot.toastit.feature.climatecocktail.dto.AreaRequestDTO;
-import alcoholboot.toastit.feature.climatecocktail.entity.WeatherEntity;
+import alcoholboot.toastit.feature.weathercocktail.dto.AreaRequestDTO;
+import alcoholboot.toastit.feature.weathercocktail.entity.WeatherEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
