+++ src/main/java/alcoholboot/toastit/feature/weathercocktail/service/WeatherCocktailService.java
@@ -1,7 +1,7 @@
-package alcoholboot.toastit.feature.climatecocktail.service;
+package alcoholboot.toastit.feature.weathercocktail.service;
 
-import alcoholboot.toastit.feature.climatecocktail.entity.RecommendByWeatherEntity;
-import alcoholboot.toastit.feature.climatecocktail.repository.RecommendByWeatherRepository;
+import alcoholboot.toastit.feature.weathercocktail.entity.WeatherCocktailEntity;
+import alcoholboot.toastit.feature.weathercocktail.repository.WeatherCocktailRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -10,11 +10,11 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class RecommendByWeatherService {
-    private final RecommendByWeatherRepository recommendByWeatherRepository;
+public class WeatherCocktailService {
+    private final WeatherCocktailRepository weatherCocktailRepository;
 
     public List<String> getIngredientsByWeather(String weather) {
-        RecommendByWeatherEntity entity = recommendByWeatherRepository.findByWeather(weather);
+        WeatherCocktailEntity entity = weatherCocktailRepository.findByWeather(weather);
         List<String> ingredients = new ArrayList<>();
 
         ingredients.add(entity.getIngredient1());
