+++ src/main/java/alcoholboot/toastit/feature/weathercocktail/service/WeatherService.java
@@ -1,8 +1,8 @@
-package alcoholboot.toastit.feature.climatecocktail.service;
+package alcoholboot.toastit.feature.weathercocktail.service;
 
-import alcoholboot.toastit.feature.climatecocktail.dto.*;
-import alcoholboot.toastit.feature.climatecocktail.entity.WeatherEntity;
-import alcoholboot.toastit.feature.climatecocktail.repository.WeatherRepository;
+import alcoholboot.toastit.feature.weathercocktail.dto.*;
+import alcoholboot.toastit.feature.weathercocktail.entity.WeatherEntity;
+import alcoholboot.toastit.feature.weathercocktail.repository.WeatherRepository;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
